Updating to use new sdl library

+++ src/game/units/physics.rs
@@ -1,4 +1,4 @@
-use std::ops::{Deref, Neg, Add, Sub, Mul, Div};
+use std::ops::{Neg, Add, Sub, Mul, Div};
 
 use std::f64;
 use super::drawing::{Game}; 
@@ -11,7 +11,7 @@ pub trait AsFloat {
 /// Millis represents a length of time in milliseconds as a signed integer.
 /// (NOTE: As `Millis` supports basic arithmetic: "negative time" is possible.)
 #[derive(Eq,Ord,PartialEq,PartialOrd,Clone,Copy)]
-pub struct Millis(pub i64);
+pub struct Millis(pub u128);
 
 impl Add<Millis> for Millis {
 	type Output = Millis;
